tweaks: feature slicing 1
