+++ src/app/[locale]/pricing/page.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import Navbar from "@/components/ui/Navbar";
 import Footer from "@/components/ui/Footer";
-import PricingCalculator from "./pricing-calculator";
+import { PricingCalculator } from "@/features/pricing";
 import { generateMetadata as genMeta, BASE_URL, KEYWORD_SETS } from "~/lib/seo";
 
 export const metadata: Metadata = genMeta({
