+++ src/features/leads/index.ts
@@ -1,0 +1,2 @@
+export { default as ContactForm } from "./components/ContactForm";
+export { contactSchema } from "./model/contactSchema";
