+++ src/features/pricing/index.ts
@@ -1,0 +1,2 @@
+export { default as PricingCalculator } from "./components/PricingCalculator";
+export type { Selections } from "./model/types";
