+++ src/shared/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Navbar } from "./Navbar";
+export { default as Footer } from "./Footer";
+export { default as CTABand } from "./CTABand";
