+++ src/shared/ui/CodeBlock.tsx
@@ -1,0 +1,44 @@
+"use client";
+
+import { PrismLight as SyntaxHighlighter } from "react-syntax-highlighter";
+import { vscDarkPlus } from "react-syntax-highlighter/dist/esm/styles/prism";
+import tsx from "react-syntax-highlighter/dist/esm/languages/prism/tsx";
+import typescript from "react-syntax-highlighter/dist/esm/languages/prism/typescript";
+import javascript from "react-syntax-highlighter/dist/esm/languages/prism/javascript";
+import jsx from "react-syntax-highlighter/dist/esm/languages/prism/jsx";
+import bash from "react-syntax-highlighter/dist/esm/languages/prism/bash";
+import json from "react-syntax-highlighter/dist/esm/languages/prism/json";
+import css from "react-syntax-highlighter/dist/esm/languages/prism/css";
+
+SyntaxHighlighter.registerLanguage("tsx", tsx);
+SyntaxHighlighter.registerLanguage("typescript", typescript);
+SyntaxHighlighter.registerLanguage("ts", typescript);
+SyntaxHighlighter.registerLanguage("javascript", javascript);
+SyntaxHighlighter.registerLanguage("js", javascript);
+SyntaxHighlighter.registerLanguage("jsx", jsx);
+SyntaxHighlighter.registerLanguage("bash", bash);
+SyntaxHighlighter.registerLanguage("shell", bash);
+SyntaxHighlighter.registerLanguage("json", json);
+SyntaxHighlighter.registerLanguage("css", css);
+
+interface CodeBlockProps {
+	language?: string;
+	children: string;
+}
+
+export function CodeBlock({ language, children }: CodeBlockProps) {
+	return (
+		<SyntaxHighlighter
+			style={vscDarkPlus}
+			language={language || "text"}
+			PreTag="div"
+			customStyle={{
+				margin: 0,
+				borderRadius: "0.5rem",
+				fontSize: "0.875rem",
+			}}
+		>
+			{children}
+		</SyntaxHighlighter>
+	);
+}
