+++ src/shared/ui/index.ts
@@ -1,0 +1,8 @@
+export { default as Button } from "./Button";
+export { default as SectionHeader } from "./SectionHeader";
+export { default as ServiceCard } from "./ServiceCard";
+export { default as IndustryCard } from "./IndustryCard";
+export { default as FAQItem } from "./FAQItem";
+export * from "./BrandIcons";
+export * from "./carousel";
+export { CodeBlock } from "./CodeBlock";
